split 2d render funcs into own file, added arg safeguards
